+++ app.js
@@ -30,8 +30,8 @@ const cooldowns = new Map();
 
 // Describes if the system saves the commands into the docs/.../commands.json file
 // Es preferible que es tingui a FALSE a no ser que es vulgui guardar especificament
-const wantToSaveCommands = false;
-const testing = false; // TESTING BOT
+const wantToSaveCommands = (process.env.SAVE_COMMANDS === 'true');
+const testing = (process.env.TESTING === 'true'); // TESTING BOT
 
 if (testing) {
     console.log(remove("AVÍS: TESTING ACTIVAT!"));